lorem ipsum funcionando de 10

+++ src/Utils/loremIpsum.h
@@ -2,8 +2,11 @@
 #define LOREMIPSUM_H
 
 #include <fstream>
-#include <string.h>
 #include <string>
+#include <iostream>
+#include <stdlib.h>
+#include <thread>
+#include <time.h>
 #include "../model/client.h"
 
 using namespace std;
@@ -12,14 +15,14 @@ class LoremIpsum{
 
 private:
 
-	int freq;
-	int cant_max;
+	long freq;
+	long cant_max;
 	char* path;
 	Client* client;
 
 public:
 
-	LoremIpsum(int freq, int cant_max, char* path, Client* client);
+	LoremIpsum(long freq, long cant_max, char* path, Client* client);
 
 	void iniciar();
